+++ src/features/Admin/Podcast/ViewPodcast.jsx
@@ -55,28 +55,33 @@ const Viewpodcast = (props) => {
     const [filterText, setFilterText] = React.useState('');
     const [resetPaginationToggle, setResetPaginationToggle] = React.useState(false);
     const context = useContext(MainContext);
+    const [flag, setFlag] = useState(false);
     const [data, setData] = useState([]);
 
     useEffect(() => {
         // console.log(context.a);
         getData();
-    }, []);
+    }, [flag]);
 
     const getData = async () => {
         const ans = await context.getPodcast();
         console.log(ans);
-        setData(ans.data);
+        setData(ans.data.sort((a, b) => parseFloat(b.id) - parseFloat(a.id)));
     };
 
     const deleteData = async (id) => {
-        console.log(id);
-        const ans = await context.deletePodcast(id);
-        console.log(ans);
-        if (ans.status) {
-            props.showAlert(true);
-        }
-        else {
-            props.showAlert(false);
+        let c = window.confirm('Do you really want to delete?')
+        if (c) {
+            console.log(id);
+            const ans = await context.deletePodcast(id);
+            console.log(ans);
+            if (ans.status) {
+                props.showAlert(true);
+                setFlag(!flag);
+            }
+            else {
+                props.showAlert(false);
+            }
         }
     };
 
@@ -84,31 +89,35 @@ const Viewpodcast = (props) => {
         {
             name: <h3>ID</h3>,
             selector: row => row.id,
-            sortable: true
+            sortable: true,
+            grow: 1
         },
         {
             name: <h3>Title</h3>,
-            selector: row => row.name,
-            sortable: true
+            selector: row => <NavLink style={{ color: "black", textDecoration: "none" }} to={`/viewPodcastDetails/${row.id}`}>{row.name}</NavLink>,
+            sortable: true,
+            grow: 1
         },
         {
             name: <h3>Last Modified</h3>,
-            selector: row => row.updatedAt,
-            sortable: true
+            selector: row => new Date(row.updatedAt).toLocaleDateString(),
+            sortable: true,
+            grow: 1
         },
         {
             name: <h3>Status</h3>,
             selector: row => row.status,
-            sortable: true
+            sortable: true,
+            grow: 1
         },
         {
-            name: <h3>Menu</h3>,
+            name: <h3>Actions</h3>,
             cell: (e) => {
                 return (
                     <div className="row">
                         <div style={{ marginRight: "2px", cursor: "pointer" }}>
                             <NavLink to={`/edit-podcast/${e.id}`}>
-                                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pencil-square" viewBox="0 0 16 16">
+                                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-pencil-square" viewBox="0 0 16 16">
                                     <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z" />
                                     <path fill-rule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z" />
                                 </svg>
@@ -117,13 +126,14 @@ const Viewpodcast = (props) => {
                         <div onClick={() => {
                             deleteData(e.id);
                         }} style={{ marginLeft: "2px", cursor: "pointer" }}>
-                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash-fill" viewBox="0 0 16 16">
+                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-trash-fill" viewBox="0 0 16 16">
                                 <path d="M2.5 1a1 1 0 0 0-1 1v1a1 1 0 0 0 1 1H3v9a2 2 0 0 0 2 2h6a2 2 0 0 0 2-2V4h.5a1 1 0 0 0 1-1V2a1 1 0 0 0-1-1H10a1 1 0 0 0-1-1H7a1 1 0 0 0-1 1H2.5zm3 4a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7a.5.5 0 0 1 .5-.5zM8 5a.5.5 0 0 1 .5.5v7a.5.5 0 0 1-1 0v-7A.5.5 0 0 1 8 5zm3 .5v7a.5.5 0 0 1-1 0v-7a.5.5 0 0 1 1 0z" />
                             </svg>
                         </div>
                     </div>
                 )
-            }
+            },
+            grow: 0.3
         }
     ];
 
@@ -148,6 +158,11 @@ const Viewpodcast = (props) => {
         <>
             <div>
                 <h1>View Podcasts</h1>
+                <NavLink to="/create-podcast" style={{ textDecoration: "none" }}>
+                    <Button variant="contained">
+                        Create Podcast
+                    </Button>
+                </NavLink>
                 <div>
                     <DataTable
                         columns={columns}
